Added initial code files
This is all the work I've put in so far.
Forgive the sloppy comments and such,
and I have yet to add file descriptions,
but the code does mostly run. I'll add
a few test images in the next commit.
The code doesn't run to completion, and
it's still very slow, so any debugging help
would be appreciated.