Added a Child Display, minor bug fixes
Also, it should be noted that Master and
the outside GUI have been altered to
help make GUI displaying easier. My goal
is that any sub-class can use the GUI-
provided, Static-Master-stored delegate
to register a new form into the main GUI's
thread. So far, it doesn't really work, but
I'm working on it.