Minor performance improvement
Added a Failed list to Image Sec Factory's
scan procedure, so that should prevent duplicate
testing. Also, I'll be playing with the master
color tolerance in the Constants file, trying to
figure out what works well. Later, I'll add a
control to the GUI that controls this, but for
now working with the default is valid.